Extract in-operator compiler fixtures
